Reapply "Reapply "split the schema into several files""
This reverts commit ecee130b7c0b667f7b1d1c538f33fb8abf985dfd.
