+++ postgres.yml
@@ -1,0 +1,14 @@
+# Note: This should only be used for local development. "password" is horribly insecure. If i can access this
+# on someone else's instance, i am going to crash tf out
+services:
+  db:
+    image: postgres
+    restart: always
+    ports:
+      - "8081:5432"
+    environment:
+      POSTGRES_PASSWORD: password
+    volumes:
+      - pgdata:/var/lib/postgresql/data
+volumes:
+  pgdata:
